First iteration of math refactor
