+++ src/pygama/math/_distributions/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package documentation
+"""
