+++ src/pygama/math/utils.py
@@ -1,5 +1,5 @@
 """
-pygama convenience functions.
+pygama utility functions.
 """
 import sys
 
@@ -28,23 +28,6 @@ def get_dataset_from_cmdline(args, run_db, cal_db):
         ds = DataSet(run=int(args["run"][0]), md=run_db, cal=cal_db,
                      v=args["verbose"])
     return ds
-
-
-def sh(cmd, sh=False):
-    """
-    input a shell command as you would type it on the command line.
-    """
-    decoders = []
-    for sub in DataTaker.__subclasses__():
-        for subsub in sub.__subclasses__():
-            try:
-                decoder = subsub(object_info) # initialize the decoder
-                decoders.append(decoder)
-            except Exception as e:
-                print(e)
-                pass
-    return decoders
-
 
 
 def tqdm_range(start, stop, step=1, verbose=False, text=None, bar_length=20, unit=None):
@@ -86,7 +69,6 @@ def tqdm_range(start, stop, step=1, verbose=False, text=None, bar_length=20, uni
     iterable : tqdm.trange
         object that can be iterated over in a for loop
     """
-
     hide_bar = True
     if isinstance(verbose, int):
         if verbose > 0:
@@ -208,112 +190,3 @@ def tree_draw(tree, vars, tcut):
             np_arrs[i + j] = np.array([tmp[k] for k in range(n)])
 
     return tuple(np_arrs)
-
-
-def peakdet(v, delta, x=None):
-    """
-    Converted from MATLAB script at: http://billauer.co.il/peakdet.html
-    Returns two arrays: [maxtab, mintab] = peakdet(v, delta, x)
-    An updated (vectorized) version is in pygama.dsp.transforms.peakdet
-    """
-    maxtab, mintab = [], []
-
-    if x is None:
-        x = np.arange(len(v))
-    v = np.asarray(v)
-
-    # sanity checks
-    if len(v) != len(x):
-        exit("Input vectors v and x must have same length")
-    if not np.isscalar(delta):
-        exit("Input argument delta must be a scalar")
-    if delta <= 0:
-        exit("Input argument delta must be positive")
-
-    maxes, mins = [], []
-    min, max = np.inf, -np.inf
-    find_max = True
-    for i in range(len(x)):
-
-        # for i=0, all 4 of these get set
-        if v[i] > max:
-            max, imax = v[i], x[i]
-        if v[i] < min:
-            min, imin = v[i], x[i]
-
-        if find_max:
-            # if the sample is less than the current max,
-            # declare the previous one a maximum, then set this as the new "min"
-            if v[i] < max - delta:
-                maxes.append((imax, max))
-                min, imin = v[i], x[i]
-                find_max = False
-        else:
-            # if the sample is more than the current min,
-            # declare the previous one a minimum, then set this as the new "max"
-            if v[i] > min + delta:
-                mins.append((imin, min))
-                max, imax = v[i], x[i]
-                find_max = True
-
-    return np.array(maxes), np.array(mins)
-
-
-def linear_fit_by_sums(x, y, var=1):
-    """
-    Fast computation of weighted linear least squares fit to a linear model
-
-    Note: doesn't compute covariances. If you want covariances, just use polyfit
-
-    Parameters
-    ----------
-    x : array like
-        x values for the fit
-    y : array like
-        y values for the fit
-    var : array like (optional)
-        The variances for each y-value
-
-    Returns
-    -------
-    (m, b) : tuple (float, float)
-        The slope (m) and y-intercept (b) of the best fit (in the least-squares
-        sense) of the data to y = mx + b
-    """
-    y = y/var
-    x = x/var
-    sum_wts = len(y)/var if np.isscalar(var) else np.sum(1/var)
-    sum_x = np.sum(x)
-    sum_xx = np.sum(x*x)
-    sum_y = np.sum(y)
-    sum_yx = np.sum(y*x)
-    m = (sum_wts * sum_yx - sum_y * sum_x) / (sum_wts * sum_xx - sum_x**2)
-    b = (sum_y - m * sum_x) / sum_wts
-    return m, b
-
-
-def fit_simple_scaling(x, y, var=1):
-    """
-    Fast computation of weighted linear least squares fit to a simple scaling
-
-    I.e. y = scale * x. Returns the best fit scale parameter and its variance.
-
-    Parameters
-    ----------
-    x : array like
-        x values for the fit
-    y : array like
-        y values for the fit
-    var : array like (optional)
-        The variances for each y-value
-
-    Returns
-    -------
-    scale, scale_var: tuple (float, float)
-        The scale parameter and its variance
-    """
-    x = np.asarray(x)
-    y = np.asarray(y)
-    scale_var = 1/np.sum(x*x/var)
-    scale = np.sum(y*x/var) * scale_var
-    return scale, scale_var
